+++ changelog.md
@@ -9,7 +9,8 @@
     - If your revolver has the "Noise" perk with a positive value, the sensors will treat it like an explosion instead
 - Add recipes for new 1,19 items (BluSunrize)
     - Mangroves can go in the sawmill
-    - Mud can be made in the bottling machine 
+    - Mud can be made in the bottling machine
+    - Moss can be grown on cobblestone in the cloche
 - Fix rendering issues with Railgun, Chemthrower and Shield (Malte)
 - Fix IO buttons in the circuit table (Malte)
 - Fix conveyor models to work properly with stained glass, eliminate clipping and fix depthbuffer issues (BluSunrize)